requirements.txt
fastapi
uvicorn
transformers
peft
sentence-transformers
faiss-cpu
torch
streamlit
requests
langchain
langgraph
openai